backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-allauth==0.63.6
django-cors-headers==4.4.0
django-encrypted-model-fields==0.6.5
cryptography==42.0.8
google-auth==2.30.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gspread==6.1.2
google-api-python-client==2.134.0
djangorestframework-simplejwt==5.3.1
Pillow>=11.0.0
python-decouple==3.8
